Support StatusException in CoroutineContextServerInterceptor. (#249)
* Support StatusException in CoroutineContextServerInterceptor.

Previously, there was no way for a subclass of CoroutineContextServerInterceptor to close a call and prevent further processing. The interceptor would unconditionally invoke the next handler in the chain unless prevented by an exception. If the coroutineContext method threw an exception, it would bubble up to the thread pool and show up in the logs as an error.

This change adds a handler for StatusException in CoroutineContextServerInterceptor to catch a StatusException and close the call with no further action.

Fixes #221

* Rethrow the exception rather than returning an empty listener on StatusException

Co-authored-by: James Ward <james@jamesward.com>