Allow kt_jvm_proto_library to have java_proto_library dep.
Using the same argument naming as in `kt_jvm_grpc_library`. Not touching
that rule, as it already depend on a `java_proto_library`, though still
not the right way, as it still always do the java_grpc_library building
in addition... Not sure how to handle that rule in a similar way...

Though that library rule handles the build-caching better it seems, so
less of an issue.