Add kt_jvm_proto_helper as individual rule. (#396)
* Add kt_jvm_proto_helper as individual rule.

This allows the BUILD file to have separate java proto and kt proto
rules in case there is a mix of java and proto projects in the same
repo, and there are stricter requirement to the output (e.g. leaner
dependency chains, smaller binaries for pure-java outputs etc.).

This also bypasses an issue in the other rules where the
kt_jvm_*_library would compile the whole dependency chain to kotlin, and
if the bazel repo is large with many proto dependencies, an can easily
end up caching all of these when containing the same classes. If only
some of these are refreshed / synced on changes you can easily end up
with an IDE claiming your new fields still dont' exist, or removed
classes are still there etc.

* Also make rule for gRPC part.

Which makes btth kr_proto and kt_grpc helper functions.

* Alternate approach to support java targets

* Fix error introduced during merge

---------

Co-authored-by: James Ward <james@jamesward.com>
Co-authored-by: aragos <aragos@gmail.com>